General updates and improvements to the language guide
In reading over the language guide I thought it could use some cleanup, so this changeset contains a bunch of copy edits.  It also makes the audience being addressed consistent, though I think a separate pull request may come later to change it from the third person to the first person to match the rest of the guides.

A minor formatting fix was also made in the Python guide.